tgemu: fix 64-bit bug causing screen artifacts

+++ tgemu/src/render.c
@@ -57,7 +57,7 @@ int render_init(void)
     ptr = malloc(0x400);
     if(!ptr) return (0);
     memset(ptr, 0, 0x400);
-    xlat[0] = (uint8 *)(((uint32)ptr + 0x200) & ~0x1FF);
+    xlat[0] = (uint8 *)(((unsigned long)ptr + 0x200) & ~0x1FF);
     xlat[1] = xlat[0] + 0x100;
 
     /* Make VCE data to raw pixel look-up table */
